katrinconsole/requirements.txt
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.3
django-cors-middleware==1.5.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
josepy==1.8.0
jsonschema==3.2.0
jwcrypto==0.8
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.21.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
pandas==1.2.5
prompt-toolkit==3.0.19
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.3
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
